refactor: Add RESTHeart status command
Add a new command 'status' to the 'rh.js' script that shows the status of RESTHeart. This command can be invoked with the '-p' or '--port' option to specify the HTTP port. The status is displayed using colored output, with green indicating that RESTHeart is running at the specified port, and yellow indicating that RESTHeart is not running.